+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Support optional callback arg for `ActiveCampaign.Contact.bulk_import/2`
 
+### Fixes
+
+* Fix `ActiveCampaign.Contact.get/1`. It was previously returning the same results from `ActiveCampaign.Contact.get_data/1`
+
 ## 0.2.1
 
 ### Enhancements